# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Billing & Inventory Management System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type AuthUserRole = typeof AuthUserRole[keyof typeof AuthUserRole];


export const AuthUserRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface AuthUser {
  id: number;
  name: string;
  username: string;
  /** @nullable */
  email?: string | null;
  role: AuthUserRole;
  isActive?: boolean;
}

export interface AuthResponse {
  user: AuthUser;
}

export interface MessageResponse {
  message: string;
}

export interface AdminDashboard {
  totalProducts: number;
  totalInventory: number;
  totalOrders: number;
  todaySales: number;
  monthlySales: number;
  totalUsers: number;
  outOfStockCount: number;
  lowStockCount: number;
}

export interface UserDashboard {
  ordersToday: number;
  totalOrders: number;
  totalRevenue: number;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  availableQuantity: number;
  createdAt: string;
  updatedAt: string;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  price: number;
  availableQuantity: number;
}

export interface ProductUpdate {
  name?: string;
  description?: string;
  price?: number;
  availableQuantity?: number;
}

export interface ProductListResponse {
  data: Product[];
  total: number;
  page: number;
  limit: number;
}

export type InventoryTransactionActionType = typeof InventoryTransactionActionType[keyof typeof InventoryTransactionActionType];


export const InventoryTransactionActionType = {
  add: 'add',
  reduce: 'reduce',
  order: 'order',
} as const;

export interface InventoryTransaction {
  id: number;
  productId: number;
  productName: string;
  previousQuantity: number;
  /** @nullable */
  quantityAdded?: number | null;
  /** @nullable */
  quantityReduced?: number | null;
  currentQuantity: number;
  actionType: InventoryTransactionActionType;
  /** @nullable */
  updatedById?: number | null;
  /** @nullable */
  updatedByName?: string | null;
  updatedAt: string;
}

export type InventoryAdjustInputActionType = typeof InventoryAdjustInputActionType[keyof typeof InventoryAdjustInputActionType];


export const InventoryAdjustInputActionType = {
  add: 'add',
  reduce: 'reduce',
} as const;

export interface InventoryAdjustInput {
  productId: number;
  quantity: number;
  actionType: InventoryAdjustInputActionType;
}

export interface InventoryListResponse {
  data: InventoryTransaction[];
  total: number;
  page: number;
  limit: number;
}

export interface Customer {
  id: number;
  name: string;
  mobile: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  gstNumber?: string | null;
  createdAt: string;
}

export interface CustomerInput {
  /** @minLength 1 */
  name: string;
  mobile: string;
  email?: string;
  address?: string;
  gstNumber?: string;
}

export interface CustomerUpdate {
  name?: string;
  mobile?: string;
  email?: string;
  address?: string;
  gstNumber?: string;
}

export interface CustomerListResponse {
  data: Customer[];
  total: number;
  page: number;
  limit: number;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  quantity: number;
  unitPrice: number;
  total: number;
}

export interface OrderItemInput {
  productId: number;
  quantity: number;
  unitPrice: number;
}

export interface Order {
  id: number;
  invoiceNumber: string;
  /** @nullable */
  customerId?: number | null;
  customerName: string;
  customerMobile: string;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  customerAddress?: string | null;
  subtotal: number;
  gstPercentage: number;
  gstAmount: number;
  referralCharges?: number;
  discount?: number;
  grandTotal: number;
  status: string;
  /** @nullable */
  createdById?: number | null;
  /** @nullable */
  createdByName?: string | null;
  createdAt: string;
}

export interface OrderDetail {
  id: number;
  invoiceNumber: string;
  /** @nullable */
  customerId?: number | null;
  customerName: string;
  customerMobile: string;
  /** @nullable */
  customerEmail?: string | null;
  /** @nullable */
  customerAddress?: string | null;
  subtotal: number;
  gstPercentage: number;
  gstAmount: number;
  referralCharges?: number;
  discount?: number;
  grandTotal: number;
  status: string;
  /** @nullable */
  createdById?: number | null;
  /** @nullable */
  createdByName?: string | null;
  createdAt: string;
  items: OrderItem[];
}

export interface OrderInput {
  customerId?: number;
  /** @minLength 1 */
  customerName: string;
  customerMobile: string;
  customerEmail?: string;
  customerAddress?: string;
  /** @minItems 1 */
  items: OrderItemInput[];
  gstPercentage: number;
  referralCharges?: number;
  discount?: number;
}

export interface OrderListResponse {
  data: Order[];
  total: number;
  page: number;
  limit: number;
}

export type AppUserRole = typeof AppUserRole[keyof typeof AppUserRole];


export const AppUserRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface AppUser {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  username: string;
  role: AppUserRole;
  isActive: boolean;
  createdAt: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface UserInput {
  /** @minLength 1 */
  name: string;
  email?: string;
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
  role: UserInputRole;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface UserUpdate {
  name?: string;
  email?: string;
  username?: string;
  role?: UserUpdateRole;
  isActive?: boolean;
  /** @minLength 6 */
  password?: string;
}

export interface ResetPasswordInput {
  /** @minLength 6 */
  newPassword: string;
}

export type ListProductsParams = {
search?: string;
page?: number;
limit?: number;
};

export type ListInventoryParams = {
productId?: number;
page?: number;
limit?: number;
};

export type ListCustomersParams = {
search?: string;
page?: number;
limit?: number;
};

export type ListOrdersParams = {
search?: string;
customerId?: number;
dateFrom?: string;
dateTo?: string;
page?: number;
limit?: number;
};