# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Billing & Inventory Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List inventory transactions
      parameters:
        - name: productId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Inventory transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryListResponse"
    post:
      operationId: adjustInventory
      tags: [inventory]
      summary: Add or reduce inventory for a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryAdjustInput"
      responses:
        "201":
          description: Inventory adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryTransaction"

  # ── Customers ─────────────────────────────────────────────────────────────
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      summary: List all customers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    patch:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Orders ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/AuthUser"
    AuthUser:
      type: object
      required: [id, name, username, role]
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        email:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, user]
        isActive:
          type: boolean
...
    InventoryTransaction:
      type: object
      required: [id, productId, productName, previousQuantity, currentQuantity, actionType, updatedAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        previousQuantity:
          type: integer
        quantityAdded:
          type: ["integer", "null"]
        quantityReduced:
          type: ["integer", "null"]
        currentQuantity:
          type: integer
        actionType:
          type: string
          enum: [add, reduce, order]
        updatedById:
          type: ["integer", "null"]
        updatedByName:
          type: ["string", "null"]
        updatedAt:
          type: string
    InventoryAdjustInput:
      type: object
      required: [productId, quantity, actionType]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        actionType:
          type: string
          enum: [add, reduce]
    InventoryListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/InventoryTransaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── Customers ─────────────────────────────────────────────────────────
    Customer:
      type: object
      required: [id, name, mobile, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        mobile:
          type: string
        email:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        gstNumber:
          type: ["string", "null"]
        createdAt:
          type: string
    CustomerInput:
      type: object
      required: [name, mobile]
      properties:
        name:
          type: string
          minLength: 1
        mobile:
          type: string
        email:
          type: string
        address:
          type: string
        gstNumber:
          type: string
    CustomerUpdate:
      type: object
      properties:
        name:
          type: string
        mobile:
          type: string
        email:
          type: string
        address:
          type: string
        gstNumber:
          type: string
    CustomerListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── Orders ────────────────────────────────────────────────────────────
...